review: moved window event types to platform.ts
