+++ api/node/typescript/index.ts
@@ -5,12 +5,12 @@ import * as napi from "../binding.cjs";
 export {
     Diagnostic,
     DiagnosticLevel,
+    WindowEventDispatchResult,
     RgbaColor,
     Brush,
     DataTransfer,
     StyledText,
     Keys,
-    WindowEventDispatchResult,
 } from "../binding.cjs";
 
 import { Model } from "./models";
@@ -20,10 +20,12 @@ export { ArrayModel } from "./models";
 
 export { language } from "./generated/language";
 
+import * as platform from "./platform";
+export { platform };
+
 import { Diagnostic } from "../binding.cjs";
 
 import { fileURLToPath } from "node:url";
-import type { language } from "./generated/language";
 
 /**
  *  Represents a two-dimensional point.
@@ -54,109 +56,6 @@ export interface Size {
      */
     height: number;
 }
-
-/** A pointer was pressed. */
-export interface PointerPressedEvent {
-    type: "pointer-pressed";
-    position: Point;
-    /** The button that was pressed. */
-    button: language.PointerEventButton;
-}
-
-/** A pointer was released. */
-export interface PointerReleasedEvent {
-    type: "pointer-released";
-    position: Point;
-    /** The button that was released. */
-    button: language.PointerEventButton;
-}
-
-/** The position of the pointer has changed. */
-export interface PointerMovedEvent {
-    type: "pointer-moved";
-    position: Point;
-}
-
-/** The wheel button of a mouse was rotated to initiate scrolling. */
-export interface PointerScrolledEvent {
-    type: "pointer-scrolled";
-    /** The position of the pointer when the scroll occurred. */
-    position: Point;
-    /** The amount of logical pixels to scroll in the horizontal direction. */
-    deltaX: number;
-    /** The amount of logical pixels to scroll in the vertical direction. */
-    deltaY: number;
-}
-
-/** The pointer exited the window. */
-export interface PointerExitedEvent {
-    type: "pointer-exited";
-}
-
-/** A key was pressed. */
-export interface KeyPressedEvent {
-    type: "key-pressed";
-    /** The unicode representation of the key pressed. */
-    text: string;
-}
-
-/** A key press was auto-repeated. */
-export interface KeyPressRepeatedEvent {
-    type: "key-press-repeated";
-    /** The unicode representation of the key pressed. */
-    text: string;
-}
-
-/** A key was released. */
-export interface KeyReleasedEvent {
-    type: "key-released";
-    /** The unicode representation of the key released. */
-    text: string;
-}
-
-/**
- * The window's scale factor has changed. This can happen for example when the display's resolution
- *  changes, the user selects a new scale factor in the system settings, or the window is moved to a
- * different screen.
- */
-export interface ScaleFactorChangedEvent {
-    type: "scale-factor-changed";
-    /** The window system provided scale factor to map logical pixels to physical pixels. */
-    scaleFactor: number;
-}
-
-/** The window was resized. */
-export interface ResizedEvent {
-    type: "resized";
-    /** The new logical size of the window */
-    size: Size;
-}
-
-/** The user requested to close the window. */
-export interface CloseRequestedEvent {
-    type: "close-requested";
-}
-
-/** The Window was activated or de-activated. */
-export interface WindowActiveChangedEvent {
-    type: "window-active-changed";
-    active: boolean;
-}
-
-/** A event that describes user input or windowing system events. */
-export type WindowEvent =
-    | PointerPressedEvent
-    | PointerReleasedEvent
-    | PointerMovedEvent
-    | PointerScrolledEvent
-    | PointerExitedEvent
-    | KeyPressedEvent
-    | KeyPressRepeatedEvent
-    | KeyReleasedEvent
-    | ScaleFactorChangedEvent
-    | ResizedEvent
-    | CloseRequestedEvent
-    | WindowActiveChangedEvent;
 
 /**
  * This type represents a window towards the windowing system, that's used to render the
@@ -204,7 +103,7 @@ export interface Window {
     requestRedraw(): void;
 
     /** Dispatch a window event to the scene. **/
-    dispatchEvent(event: WindowEvent): napi.WindowEventDispatchResult;
+    dispatchEvent(event: platform.WindowEvent): napi.WindowEventDispatchResult;
 }
 
 /**
