+++ knip.json
@@ -8,7 +8,8 @@
       "ignoreFiles": ["editors/tree-sitter-slint/grammar.js"]
     },
     "api/node": {
-      "project": ["typescript/**/*.ts", "__test__/**/*.{ts,mts}"]
+      "project": ["typescript/**/*.ts", "__test__/**/*.{ts,mts}"],
+      "ignoreExportsUsedInFile": true
     },
     "editors/vscode": {
       "entry": ["src/extension.ts", "src/browser.ts", "src/browser-lsp-worker.ts"],
